input 컴포넌트화
